move types to local types
